requirements.txt
crewai
crewai-tools
python-dotenv
beautifulsoup4
requests
yfinance
openai
langchain-groq
langchain-community